Refactor the Segment class
This commit aims at simplifying the Segment class. Here are the major changes:
  - instead of referring to steps, samples in a segment are referred to
  as samples (i.e. certain variables and methods were renamed);
  - every segment allows it's parameters (e.g. length, amplitude) to be modified
  during the lifetime of the underlying object;
  - the user of the class is responsible for regeneration of the samples
  whenever the parameters of the segment are updated;
  - one segment class can be used for generating samples for various
  configurations (instead of creating a new object when, for example,
  updating the amplitude);
The segment family of classes still assumes that the Constant, Linear
and Exponential segments are separate Classes (inheriting from the
same base class). However, now there's more code sharing withing this
hierarchy.

The changes above enabled refactoring of the tests, so that more code
can be re-used between the classes. In particular, I've introduced test
fixtures (GTest nomenclature), so that the code is cleaner.